Added css files
Each of you have a personal css file: Write your code inside there,
link your css file to the html file and you should see changes in
styling.
I Remember to you all that this is for the milestone.
If you do not remember what to do exactly, review the slides and
everything will be clear.